requirements.txt
discord.py==2.3.1
python-dotenv==1.0.0
validators==0.20.0
yt-dlp>=2026.3.17
pyyaml==6.0.1